Snyk bundle/2025 09 09 (#58)
* fix: upgrade @radix-ui/react-accordion from 1.2.3 to 1.2.8

Snyk has created this PR to upgrade @radix-ui/react-accordion from 1.2.3 to 1.2.8.

See this package in npm:
@radix-ui/react-accordion

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade @radix-ui/react-tabs from 1.1.3 to 1.1.11 (#38)

Snyk has created this PR to upgrade @radix-ui/react-tabs from 1.1.3 to 1.1.11.

See this package in npm:
@radix-ui/react-tabs

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @radix-ui/react-checkbox from 1.1.4 to 1.3.1 (#39)

Snyk has created this PR to upgrade @radix-ui/react-checkbox from 1.1.4 to 1.3.1.

See this package in npm:
@radix-ui/react-checkbox

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade next from 15.3.1 to 15.4.6 (#55)

Snyk has created this PR to upgrade next from 15.3.1 to 15.4.6.

See this package in npm:
next

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade zod from 3.24.2 to 3.25.76 (#47)

Snyk has created this PR to upgrade zod from 3.24.2 to 3.25.76.

See this package in npm:
zod

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade framer-motion from 12.6.3 to 12.23.12 (#54)

Snyk has created this PR to upgrade framer-motion from 12.6.3 to 12.23.12.

See this package in npm:
framer-motion

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @radix-ui/react-radio-group from 1.2.3 to 1.3.7 (#45)

Snyk has created this PR to upgrade @radix-ui/react-radio-group from 1.2.3 to 1.3.7.

See this package in npm:
@radix-ui/react-radio-group

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade isomorphic-dompurify from 2.22.0 to 2.25.0 (#46)

Snyk has created this PR to upgrade isomorphic-dompurify from 2.22.0 to 2.25.0.

See this package in npm:
isomorphic-dompurify

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @radix-ui/react-dropdown-menu from 2.1.6 to 2.1.16 (#56)

Snyk has created this PR to upgrade @radix-ui/react-dropdown-menu from 2.1.6 to 2.1.16.

See this package in npm:
@radix-ui/react-dropdown-menu

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @radix-ui/react-menubar from 1.1.6 to 1.1.16 (#57)

Snyk has created this PR to upgrade @radix-ui/react-menubar from 1.1.6 to 1.1.16.

See this package in npm:
@radix-ui/react-menubar

See this project in Snyk:
https://app.snyk.io/org/awesumsawz/project/fb5777b3-db1c-4cf5-8b33-2b9a4d40b3ae?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fresh npm install

---------

Co-authored-by: snyk-bot <snyk-bot@snyk.io>